% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contributions.R
\name{get_contributions}
\alias{get_contributions}
\title{Get GitHub contributions}
\source{
https://stackoverflow.com/a/75277425/9370662
}
\usage{
get_contributions(repo, user = "easystats")
}
\arguments{
\item{repo}{repository}

\item{user}{username}
}
\value{
A data frame with username of contributor, number of added lines of code,
deleted lines of code, and number of commits.
}
\description{
Get GitHub contributions, which can then be ordered with \code{order_authors()}.
}
\examples{
head(get_contributions("report"))
}
