% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_authors.R
\name{order_authors}
\alias{order_authors}
\title{Order GitHub authors}
\usage{
order_authors(data, commit.weight = 0, cutoff = 1000)
}
\arguments{
\item{data}{A data frame or a named list of data frames (as returned by
\code{get_contributions()} when multiple repositories are provided).}

\item{commit.weight}{A multiplication constant representing the weight the
number of commits should have (setting it to 0, the default, will make it
have no impact).}

\item{cutoff}{Cutoff score under which to exclude contributors from the
filtered list.}
}
\value{
When a single data frame is provided, a data frame with username of
contributor, number of added lines of code, deleted lines of code, number
of commits, and a total 'score' column representing the sum of the other
columns. When a list of data frames is provided, a named list of data
frames with the same structure.
}
\description{
Order GitHub authors based on contributions (obtained from \code{get_contributions()}).
}
\examples{
x <- get_contributions("report")
order_authors(x)
}
